pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "greenpipeline"
version = "0.1.0"
description = "AI-powered CI/CD pipeline optimization and carbon analysis system"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "GreenPipeline Team" }]
license = "MIT"
dependencies = [
    "networkx>=3.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "typing_extensions>=4.13.0",
]

[project.optional-dependencies]
cli = ["rich>=13.7.1", "typer", "questionary"]
ml = ["pandas>=2.0", "datasets"]
infra = ["docker>=7.0", "kubernetes>=30.1.0", "paramiko>=3.5.0"]
observability = [
    "opentelemetry-sdk>=1.23.0",
    "prometheus_client",
    "codecarbon>=2.3",
]
dashboard = ["streamlit>=1.28"]
dev = ["ty", "mypy", "pre-commit", "pytest", "ruff"]

[project.scripts]
greenpipeline = "greenpipeline.pipeline_runner:main"

[tool.setuptools.packages.find]
include = ["greenpipeline*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ty.environment]
python-version = "3.11"

[tool.ty.analysis]
replace-imports-with-any = [
    "networkx.**",
    "matplotlib.**",
    "streamlit.**",
    "codecarbon.**",
    "kubernetes.**",
    "docker.**",
    "pandas.**",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
